+++ contracts/interfaces/IHexagonOracle.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.9;
 
 interface IHexagonOracle {
-    function getTotalETHStakingProceeds(uint256 epoch) external view returns (uint256);
+    function getTotalETHStakingProceeds(uint32 epoch) external view returns (uint256);
 }
